+++ docker_daemon/check.py
@@ -1,4 +1,4 @@
-# (C) Datadog, Inc. 2010-2016
+# (C) Datadog, Inc. 2010-2017
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 
@@ -66,8 +66,14 @@ CGROUP_METRICS = [
             "docker.mem.sw_limit": (["hierarchical_memsw_limit"], lambda x: float(x) if float(x) < 2 ** 60 else None, GAUGE),
             "docker.mem.in_use": (["rss", "hierarchical_memory_limit"], lambda x, y: float(x)/float(y) if float(y) < 2 ** 60 else None, GAUGE),
             "docker.mem.sw_in_use": (["swap", "rss", "hierarchical_memsw_limit"], lambda x, y, z: float(x + y)/float(z) if float(z) < 2 ** 60 else None, GAUGE)
-
         }
+    },
+    {
+        "cgroup": "memory",
+        "file": "memory.soft_limit_in_bytes",
+        "metrics": {
+            "softlimit": ("docker.mem.soft_limit", GAUGE),
+        },
     },
     {
         "cgroup": "cpuacct",
@@ -173,10 +179,15 @@ class DockerDaemon(AgentCheck):
             raise Exception("Docker check only supports one configured instance.")
         AgentCheck.__init__(self, name, init_config,
                             agentConfig, instances=instances)
-
         self.init_success = False
         self._service_discovery = agentConfig.get('service_discovery') and \
             agentConfig.get('service_discovery_backend') == 'docker'
+
+        global_labels_as_tags = agentConfig.get('docker_labels_as_tags')
+        if global_labels_as_tags:
+            self.collect_labels_as_tags = [label.strip() for label in global_labels_as_tags.split(',')]
+        else:
+            self.collect_labels_as_tags = DEFAULT_LABELS_AS_TAGS
         self.init()
 
     def init(self):
@@ -209,7 +220,12 @@ class DockerDaemon(AgentCheck):
             self._disable_net_metrics = False
 
             # Set tagging options
-            self.collect_labels_as_tags = instance.get("collect_labels_as_tags", DEFAULT_LABELS_AS_TAGS)
+            # The collect_labels_as_tags is legacy, only tagging docker metrics.
+            # It is replaced by docker_labels_as_tags in datadog.conf.
+            # We keep this line for backward compatibility.
+            if "collect_labels_as_tags" in instance:
+                self.collect_labels_as_tags = instance.get("collect_labels_as_tags")
+
             self.kube_pod_tags = {}
 
             self.use_histogram = _is_affirmative(instance.get('use_histogram', False))
@@ -399,13 +415,19 @@ class DockerDaemon(AgentCheck):
                 except Exception as e:
                     self.log.debug("Unable to inspect Docker container: %s", e)
 
+        total_count = 0
         # TODO: deprecate these 2, they should be replaced by _report_container_count
         for tags, count in running_containers_count.iteritems():
+            total_count += count
             self.gauge("docker.containers.running", count, tags=list(tags))
+        self.gauge("docker.containers.running.total", total_count, tags=self.custom_tags)
 
+        total_count = 0
         for tags, count in all_containers_count.iteritems():
             stopped_count = count - running_containers_count[tags]
+            total_count += stopped_count
             self.gauge("docker.containers.stopped", stopped_count, tags=list(tags))
+        self.gauge("docker.containers.stopped.total", total_count, tags=self.custom_tags)
 
         return containers_by_id
 
@@ -628,7 +650,7 @@ class DockerDaemon(AgentCheck):
     def _report_performance_metrics(self, containers_by_id):
 
         containers_without_proc_root = []
-        for container in containers_by_id.itervalues():
+        for container_id, container in containers_by_id.iteritems():
             if self._is_container_excluded(container) or not self._is_container_running(container):
                 continue
 
@@ -641,7 +663,7 @@ class DockerDaemon(AgentCheck):
                     continue
                 self._report_net_metrics(container, tags)
             except BogusPIDException as e:
-                self.log.warning('Unable to report cgroup metrics: %s', e)
+                self.log.warning('Unable to report cgroup metrics for container %s: %s', container_id[:12], e)
 
         if containers_without_proc_root:
             message = "Couldn't find pid directory for containers: {0}. They'll be missing network metrics".format(
@@ -994,6 +1016,13 @@ class DockerDaemon(AgentCheck):
                     return self._parse_blkio_metrics(fp.read().splitlines())
                 elif 'cpuacct.usage' in stat_file:
                     return dict({'usage': str(int(fp.read())/10000000)})
+                elif 'memory.soft_limit_in_bytes' in stat_file:
+                    value = int(fp.read())
+                    # do not report kernel max default value (uint64 * 4096)
+                    # see https://github.com/torvalds/linux/blob/5b36577109be007a6ecf4b65b54cbc9118463c2b/mm/memcontrol.c#L2844-L2845
+                    # 2 ** 60 is kept for consistency of other cgroups metrics
+                    if value < 2 ** 60:
+                        return dict({'softlimit': value})
                 else:
                     return dict(map(lambda x: x.split(' ', 1), fp.read().splitlines()))
         except IOError: