+++ docker_daemon/test_docker_daemon.py
@@ -1,4 +1,4 @@
-# (C) Datadog, Inc. 2010-2016
+# (C) Datadog, Inc. 2010-2017
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 
@@ -13,6 +13,7 @@ from nose.plugins.attrib import attr
 # project
 from checks import AgentCheck
 from tests.checks.common import AgentCheckTest
+from tests.checks.common import load_check
 from utils.dockerutil import DockerUtil
 
 log = logging.getLogger('tests')
@@ -292,6 +293,8 @@ class TestCheckDockerDaemon(AgentCheckTest):
             ('docker.containers.running', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest']),
+            ('docker.containers.running.total', None),
+            ('docker.containers.stopped.total', None),
             ('docker.image.size', ['image_name:nginx', 'image_tag:latest']),
             ('docker.image.size', ['image_name:redis', 'image_tag:latest']),
             ('docker.image.virtual_size', ['image_name:nginx', 'image_tag:latest']),
@@ -325,6 +328,8 @@ class TestCheckDockerDaemon(AgentCheckTest):
             ('docker.containers.running', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest']),
+            ('docker.containers.running.total', None),
+            ('docker.containers.stopped.total', None),
             ('docker.images.available', None),
             ('docker.images.intermediate', None),
             ('docker.cpu.system', ['container_name:test-new-nginx-latest', 'docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest']),
@@ -372,6 +377,8 @@ class TestCheckDockerDaemon(AgentCheckTest):
             ('docker.containers.running', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
+            ('docker.containers.running.total', None),
+            ('docker.containers.stopped.total', None),
             ('docker.cpu.system', ['container_name:test-new-redis-latest', 'docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.cpu.user', ['container_name:test-new-redis-latest', 'docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.image.size', ['image_name:redis', 'image_tag:latest']),
@@ -427,6 +434,8 @@ class TestCheckDockerDaemon(AgentCheckTest):
             ('docker.containers.running', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
+            ('docker.containers.running.total', None),
+            ('docker.containers.stopped.total', None),
             ('docker.cpu.system', ['container_name:test-new-redis-latest', 'docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.cpu.user', ['container_name:test-new-redis-latest', 'docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.image.size', ['image_name:redis', 'image_tag:latest']),
@@ -482,6 +491,8 @@ class TestCheckDockerDaemon(AgentCheckTest):
             ('docker.containers.running', ['container_command:docker-entrypoint.sh redis-server']),
             ('docker.containers.stopped', ["container_command:nginx -g 'daemon off;'"]),
             ('docker.containers.stopped', ['container_command:docker-entrypoint.sh redis-server']),
+            ('docker.containers.running.total', None),
+            ('docker.containers.stopped.total', None),
             ('docker.cpu.system', ["container_command:nginx -g 'daemon off;'"]),
             ('docker.cpu.system', ['container_command:docker-entrypoint.sh redis-server']),
             ('docker.cpu.user', ['container_command:docker-entrypoint.sh redis-server']),
@@ -552,6 +563,8 @@ class TestCheckDockerDaemon(AgentCheckTest):
             ('docker.containers.running', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'label1:nginx']),
+            ('docker.containers.running.total', None),
+            ('docker.containers.stopped.total', None),
             ('docker.image.size', ['image_name:redis', 'image_tag:latest']),
             ('docker.image.size', ['image_name:nginx', 'image_tag:latest']),
             ('docker.image.virtual_size', ['image_name:redis', 'image_tag:latest']),
@@ -588,6 +601,43 @@ class TestCheckDockerDaemon(AgentCheckTest):
         for mname, tags in expected_metrics:
             self.assertMetric(mname, tags=tags, count=1, at_least=1)
 
+    def test_collect_labels_as_tags(self):
+        expected_metrics = [
+            ('docker.containers.stopped.total', None),
+            ('docker.containers.running.total', None),
+            ('docker.containers.running', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
+            ('docker.containers.running', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'label1:nginx']),
+            ('docker.mem.rss', ['container_name:test-new-nginx-latest', 'docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'label1:nginx']),
+            ('docker.containers.stopped', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
+            ('docker.containers.stopped', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'label1:nginx']),
+            ('docker.mem.rss', ['container_name:test-new-redis-latest', 'docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
+            ('docker.mem.limit', ['container_name:test-new-nginx-latest', 'docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'label1:nginx']),
+            ('docker.mem.cache', ['container_name:test-new-nginx-latest', 'docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'label1:nginx']),
+            ('docker.mem.cache', ['container_name:test-new-redis-latest', 'docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
+            ('docker.mem.in_use', ['container_name:test-new-nginx-latest', 'docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'label1:nginx']),
+        ]
+
+        config = {
+            "init_config": {},
+            "instances": [{
+                "url": "unix://var/run/docker.sock",
+            },
+            ],
+        }
+
+        DockerUtil._drop()
+        DockerUtil(init_config=config['init_config'], instance=config['instances'][0])
+
+        self.agentConfig = {
+            'docker_labels_as_tags': 'label1'
+        }
+        self.check = load_check('docker_daemon', config, self.agentConfig)
+
+        self.check.check(config)
+        self.metrics = self.check.get_metrics()
+        for mname, tags in expected_metrics:
+            self.assertMetric(mname, tags=tags, count=1, at_least=1)
+
     def test_histogram(self):
 
         metric_suffix = ["count", "avg", "median", "max", "95percentile"]
@@ -597,6 +647,8 @@ class TestCheckDockerDaemon(AgentCheckTest):
             ('docker.containers.running', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest']),
+            ('docker.containers.running.total', None),
+            ('docker.containers.stopped.total', None),
             ('docker.image.size', ['image_name:redis', 'image_tag:latest']),
             ('docker.image.size', ['image_name:nginx', 'image_tag:latest']),
             ('docker.image.virtual_size', ['image_name:nginx', 'image_tag:latest']),
@@ -688,6 +740,8 @@ class TestCheckDockerDaemon(AgentCheckTest):
             ('docker.containers.running', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:redis:latest', 'image_name:redis', 'image_tag:latest']),
             ('docker.containers.stopped', ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest']),
+            ('docker.containers.running.total', None),
+            ('docker.containers.stopped.total', None),
             ('docker.image.size', ['image_name:redis', 'image_tag:latest']),
             ('docker.image.size', ['image_name:nginx', 'image_tag:latest']),
             ('docker.image.virtual_size', ['image_name:redis', 'image_tag:latest']),
@@ -836,6 +890,10 @@ class TestCheckDockerDaemon(AgentCheckTest):
             elif 'cpu' in stat_file:
                 return {'user': 1000 * self.run, 'system': 1000 * self.run}
                 self.run += 1
+            elif 'memory.soft_limit_in_bytes' in stat_file:
+                    value = int(fp.read())
+                    if value < 2 ** 60:
+                        return dict({'softlimit': value})
             else:
                 return dict(map(lambda x: x.split(' ', 1), fp.read().splitlines()))