+++ kafka_consumer/test_kafka_consumer.py
@@ -1,4 +1,4 @@
-# (C) Datadog, Inc. 2010-2016
+# (C) Datadog, Inc. 2010-2017
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 
@@ -14,12 +14,13 @@ from nose import SkipTest
 from kafka import KafkaConsumer, KafkaProducer
 
 from kazoo.client import KazooClient
+from docker import Client  # required by test setup
 
 # project
-from tests.checks.common import AgentCheckTest
+from tests.checks.common import AgentCheckTest, log
 
 
-instances = [{
+zk_instance = {
     'kafka_connect_str': '172.17.0.1:9092',
     'zk_connect_str': 'localhost:2181',
     # 'zk_prefix': '/0.8',
@@ -28,7 +29,18 @@ instances = [{
             'marvel': [0]
         }
     }
-}]
+}
+
+kafka_instance = {
+    'kafka_connect_str': '172.17.0.1:9092',
+    'kafka_consumer_offsets': True,
+    'consumer_groups': {
+        'my_consumer': {
+            'marvel': [0]
+        }
+    }
+}
+
 
 BROKER_METRICS = [
     'kafka.broker_offset',
@@ -43,25 +55,31 @@ TOPICS = ['marvel', 'dc', '__consumer_offsets']
 PARTITIONS = [0, 1]
 
 SHUTDOWN = threading.Event()
+CLUSTER_READY = 'Stabilized group my_consumer'
+KAFKA_IMAGE_NAME = 'wurstmeister/kafka'
+DOCKER_TO = 10
 
 class Producer(threading.Thread):
 
     def run(self):
-        producer = KafkaProducer(bootstrap_servers=instances[0]['kafka_connect_str'])
+        producer = KafkaProducer(bootstrap_servers=zk_instance['kafka_connect_str'])
 
         while not SHUTDOWN.is_set():
             for partition in PARTITIONS:
-                producer.send('marvel', b"Peter Parker", partition=partition)
-                producer.send('marvel', b"Bruce Banner", partition=partition)
-                producer.send('marvel', b"Tony Stark", partition=partition)
-                producer.send('marvel', b"Johhny Blaze", partition=partition)
-                producer.send('marvel', b"\xc2BoomShakalaka", partition=partition)
-                producer.send('dc', b"Diana Prince", partition=partition)
-                producer.send('dc', b"Bruce Wayne", partition=partition)
-                producer.send('dc', b"Clark Kent", partition=partition)
-                producer.send('dc', b"Arthur Curry", partition=partition)
-                producer.send('dc', b"\xc2ShakalakaBoom", partition=partition)
-                time.sleep(1)
+                try:
+                    producer.send('marvel', b"Peter Parker", partition=partition)
+                    producer.send('marvel', b"Bruce Banner", partition=partition)
+                    producer.send('marvel', b"Tony Stark", partition=partition)
+                    producer.send('marvel', b"Johhny Blaze", partition=partition)
+                    producer.send('marvel', b"\xc2BoomShakalaka", partition=partition)
+                    producer.send('dc', b"Diana Prince", partition=partition)
+                    producer.send('dc', b"Bruce Wayne", partition=partition)
+                    producer.send('dc', b"Clark Kent", partition=partition)
+                    producer.send('dc', b"Arthur Curry", partition=partition)
+                    producer.send('dc', b"\xc2ShakalakaBoom", partition=partition)
+                    time.sleep(1)
+                except Exception:
+                    pass
 
 
 class ZKConsumer(threading.Thread):
@@ -70,7 +88,7 @@ class ZKConsumer(threading.Thread):
         zk_path_topic_tmpl = '/consumers/my_consumer/offsets/'
         zk_path_partition_tmpl = zk_path_topic_tmpl + '{topic}/{partition}'
 
-        zk_conn = KazooClient(instances[0]['zk_connect_str'], timeout=10)
+        zk_conn = KazooClient(zk_instance['zk_connect_str'], timeout=10)
         zk_conn.start()
 
         for topic in TOPICS:
@@ -81,7 +99,7 @@ class ZKConsumer(threading.Thread):
                     zk_conn.ensure_path(node_path)
                     zk_conn.set(node_path, str(0))
 
-        consumer = KafkaConsumer(bootstrap_servers=instances[0]['kafka_connect_str'],
+        consumer = KafkaConsumer(bootstrap_servers=zk_instance['kafka_connect_str'],
                                  group_id="my_consumer",
                                  auto_offset_reset='earliest',
                                  enable_auto_commit=False)
@@ -112,7 +130,7 @@ class ZKConsumer(threading.Thread):
 class KConsumer(threading.Thread):
 
     def run(self):
-        consumer = KafkaConsumer(bootstrap_servers=instances[0]['kafka_connect_str'],
+        consumer = KafkaConsumer(bootstrap_servers=kafka_instance['kafka_connect_str'],
                                  group_id="my_consumer",
                                  auto_offset_reset='earliest')
         consumer.subscribe(TOPICS)
@@ -125,6 +143,7 @@ class KConsumer(threading.Thread):
 class TestKafka(AgentCheckTest):
     """Basic Test for kafka_consumer integration."""
     CHECK_NAME = 'kafka_consumer'
+    MAX_SETUP_WAIT = 60
     THREADS = [Producer()]
 
     def __init__(self, *args, **kwargs):
@@ -137,10 +156,39 @@ class TestKafka(AgentCheckTest):
 
     @classmethod
     def setUpClass(cls):
+        """
+        Setup the consumer + producer, and wait for cluster
+        """
+        start = time.time()
+
         cls.THREADS[0].start()
         time.sleep(5)
         cls.THREADS[1].start()
         time.sleep(5)
+
+        try:
+            cli = Client(base_url='unix://var/run/docker.sock',
+                         timeout=DOCKER_TO)
+            containers = cli.containers()
+
+            nodes = []
+            for c in containers:
+                if KAFKA_IMAGE_NAME in c.get('Image'):
+                    nodes.append(c)
+
+            elapsed = time.time() - start
+            while elapsed < cls.MAX_SETUP_WAIT:
+                for node in nodes:
+                    _log = cli.logs(node.get('Id'))
+                    if CLUSTER_READY in _log:
+                        return
+
+                time.sleep(1)
+                elapsed = time.time() - start
+        except Exception:
+            pass
+
+        log.info('Unable to verify kafka cluster status - tests may fail')
 
     @classmethod
     def tearDownClass(cls):
@@ -149,14 +197,39 @@ class TestKafka(AgentCheckTest):
             if t.is_alive():
                 t.join(5)
 
+    def is_supported(self, flavors):
+        supported = False
+        version = os.environ.get('FLAVOR_VERSION')
+        flavor = os.environ.get('FLAVOR_OPTIONS','').lower()
+
+        if not version:
+            return False
+
+        for f in flavors:
+            if f == flavor:
+                supported = True
+
+        if not supported:
+            return False
+
+        if version is not 'latest':
+            version = version.split('-')[0]
+            version = tuple(s for s in version.split('.') if s.strip())
+            if flavor is 'kafka' and version <= self.check.LAST_ZKONLY_VERSION:
+                supported = False
+
+        return supported
+
+
     def test_check_zk(self):
         """
         Testing Kafka_consumer check.
         """
 
-        if os.environ.get('FLAVOR_OPTIONS','').lower() == "kafka":
-            raise SkipTest("Skipping test - environment not configured for ZK consumer offsets")
+        if not self.is_supported(['zookeeper']):
+            raise SkipTest("Skipping test - not supported in current environment")
 
+        instances = [zk_instance]
         self.run_check({'instances': instances})
 
         for instance in instances:
@@ -168,11 +241,38 @@ class TestKafka(AgentCheckTest):
                         for mname in BROKER_METRICS:
                             self.assertMetric(mname, tags=tags, at_least=1)
                         for mname in CONSUMER_METRICS:
-                            self.assertMetric(mname, tags=tags + ["consumer_group:{}".format(name)], at_least=1)
+                            self.assertMetric(mname, tags=tags + ["source:zk", "consumer_group:{}".format(name)], at_least=1)
 
         # let's reassert for the __consumer_offsets - multiple partitions
         self.assertMetric('kafka.broker_offset', at_least=1)
         self.coverage_report()
+
+    def test_multiple_servers_zk(self):
+        """
+        Testing Kafka_consumer check.
+        """
+
+        if not self.is_supported(['zookeeper']):
+            raise SkipTest("Skipping test - not supported in current environment")
+
+        multiple_server_zk_instance = copy.deepcopy(zk_instance)
+        multiple_server_zk_instance['kafka_connect_str'] = [
+            multiple_server_zk_instance['kafka_connect_str'],
+            'localhost:9092']
+
+        instances = [multiple_server_zk_instance]
+        self.run_check({'instances': instances})
+
+        for instance in instances:
+            for name, consumer_group in instance['consumer_groups'].iteritems():
+                for topic, partitions in consumer_group.iteritems():
+                    for partition in partitions:
+                        tags = ["topic:{}".format(topic),
+                                "partition:{}".format(partition)]
+                        for mname in BROKER_METRICS:
+                            self.assertMetric(mname, tags=tags, at_least=1)
+                        for mname in CONSUMER_METRICS:
+                            self.assertMetric(mname, tags=tags + ["source:zk", "consumer_group:{}".format(name)], at_least=1)
 
 
     def test_check_nogroups_zk(self):
@@ -180,16 +280,17 @@ class TestKafka(AgentCheckTest):
         Testing Kafka_consumer check grabbing groups from ZK
         """
 
-        if os.environ.get('FLAVOR_OPTIONS','').lower() == "kafka":
-            raise SkipTest("Skipping test - environment not configured for ZK consumer offsets")
+        if not self.is_supported(['zookeeper']):
+            raise SkipTest("Skipping test - not supported in current environment")
 
-        nogroup_instances = copy.deepcopy(instances)
-        nogroup_instances[0].pop('consumer_groups')
-        nogroup_instances[0]['monitor_unlisted_consumer_groups'] = True
+        nogroup_instance = copy.deepcopy(zk_instance)
+        nogroup_instance.pop('consumer_groups')
+        nogroup_instance['monitor_unlisted_consumer_groups'] = True
 
-        self.run_check({'instances': nogroup_instances})
+        instances = [nogroup_instance]
+        self.run_check({'instances': instances})
 
-        for instance in nogroup_instances:
+        for instance in instances:
             for topic in TOPICS:
                 if topic is not '__consumer_offsets':
                     for partition in PARTITIONS:
@@ -198,9 +299,35 @@ class TestKafka(AgentCheckTest):
                         for mname in BROKER_METRICS:
                             self.assertMetric(mname, tags=tags, at_least=1)
                             for mname in CONSUMER_METRICS:
-                                self.assertMetric(mname, tags=tags + ["consumer_group:my_consumer"], at_least=1)
+                                self.assertMetric(mname, tags=tags + ["source:zk", "consumer_group:my_consumer"], at_least=1)
                 else:
                     for mname in BROKER_METRICS + CONSUMER_METRICS:
                         self.assertMetric(mname, at_least=1)
 
+        self.coverage_report()
+
+    def test_check_kafka(self):
+        """
+        Testing Kafka_consumer check.
+        """
+
+        if not self.is_supported(['kafka']):
+            raise SkipTest("Skipping test - not supported in current environment")
+
+        instances = [kafka_instance]
+        self.run_check({'instances': instances})
+
+        for instance in instances:
+            for name, consumer_group in instance['consumer_groups'].iteritems():
+                for topic, partitions in consumer_group.iteritems():
+                    for partition in partitions:
+                        tags = ["topic:{}".format(topic),
+                                "partition:{}".format(partition)]
+                        for mname in BROKER_METRICS:
+                            self.assertMetric(mname, tags=tags, at_least=1)
+                        for mname in CONSUMER_METRICS:
+                            self.assertMetric(mname, tags=tags + ["source:kafka", "consumer_group:{}".format(name)], at_least=1)
+
+        # let's reassert for the __consumer_offsets - multiple partitions
+        self.assertMetric('kafka.broker_offset', at_least=1)
         self.coverage_report()