+++ snmp/test_snmp.py
@@ -1,10 +1,9 @@
-# (C) Datadog, Inc. 2010-2016
+# (C) Datadog, Inc. 2010-2017
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 
 # stdlib
 import copy
-import time
 
 # 3rd party
 from nose.plugins.attrib import attr
@@ -209,24 +208,6 @@ class SNMPTestCase(AgentCheckTest):
 
         return instance_config
 
-    def wait_for_async(self, method, attribute, count):
-        """
-        Loop on `self.check.method` until `self.check.attribute >= count`.
-
-        Raise after
-        """
-        i = 0
-        while i < RESULTS_TIMEOUT:
-            self.check._process_results()
-            if len(getattr(self.check, attribute)) >= count:
-                return getattr(self.check, method)()
-            time.sleep(1)
-            i += 1
-        raise Exception("Didn't get the right count for {attribute} in time, {total}/{expected} in {seconds}s: {attr}"
-                        .format(attribute=attribute, total=len(getattr(self.check, attribute)), expected=count, seconds=i,
-                                attr=getattr(self.check, attribute)))
-
-
     def test_command_generator(self):
         """
         Command generator's parameters should match init_config
@@ -256,7 +237,7 @@ class SNMPTestCase(AgentCheckTest):
                 self.SUPPORTED_METRIC_TYPES + self.UNSUPPORTED_METRICS)]
         }
         self.run_check_n(config, repeat=3)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for metric in self.SUPPORTED_METRIC_TYPES:
@@ -268,7 +249,7 @@ class SNMPTestCase(AgentCheckTest):
 
         # Test service check
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+                                tags=self.CHECK_TAGS, at_least=1)
 
         self.coverage_report()
 
@@ -284,15 +265,15 @@ class SNMPTestCase(AgentCheckTest):
         config = {
             'instances': [self.generate_instance_config(self.PLAY_WITH_GET_NEXT_METRICS)]
         }
-        self.run_check_twice(config)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.run_check_n(config, repeat=3)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test service check
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+                                tags=self.CHECK_TAGS, at_least=1)
 
         self.run_check(config)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for metric in self.PLAY_WITH_GET_NEXT_METRICS:
@@ -313,7 +294,7 @@ class SNMPTestCase(AgentCheckTest):
             'instances': [self.generate_instance_config(self.SCALAR_OBJECTS)]
         }
         self.run_check_n(config, repeat=3)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for metric in self.SCALAR_OBJECTS:
@@ -321,8 +302,7 @@ class SNMPTestCase(AgentCheckTest):
             self.assertMetric(metric_name, tags=self.CHECK_TAGS, count=1)
 
         # Test service check
-        self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+        self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK, tags=self.CHECK_TAGS, at_least=1)
 
         self.coverage_report()
 
@@ -334,7 +314,7 @@ class SNMPTestCase(AgentCheckTest):
             'instances': [self.generate_instance_config(self.TABULAR_OBJECTS)]
         }
         self.run_check_n(config, repeat=3, sleep=2)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for symbol in self.TABULAR_OBJECTS[0]['symbols']:
@@ -348,7 +328,7 @@ class SNMPTestCase(AgentCheckTest):
 
         # Test service check
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+                                tags=self.CHECK_TAGS, at_least=1)
 
         self.coverage_report()
 
@@ -379,7 +359,7 @@ class SNMPTestCase(AgentCheckTest):
 
 
         self.run_check_n(config, repeat=3, sleep=2)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for symbol in self.TABULAR_OBJECTS[0]['symbols']:
@@ -393,7 +373,7 @@ class SNMPTestCase(AgentCheckTest):
 
         # Test service check
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+                                tags=self.CHECK_TAGS, at_least=1)
 
         self.coverage_report()
 
@@ -424,7 +404,7 @@ class SNMPTestCase(AgentCheckTest):
 
 
         self.run_check_n(config, repeat=3, sleep=2)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for symbol in self.TABULAR_OBJECTS[0]['symbols']:
@@ -438,7 +418,7 @@ class SNMPTestCase(AgentCheckTest):
 
         # Test service check
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+                                tags=self.CHECK_TAGS, at_least=1)
 
         self.coverage_report()
 
@@ -469,7 +449,7 @@ class SNMPTestCase(AgentCheckTest):
 
 
         self.run_check_n(config, repeat=3, sleep=2)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for symbol in self.TABULAR_OBJECTS[0]['symbols']:
@@ -483,7 +463,7 @@ class SNMPTestCase(AgentCheckTest):
 
         # Test service check
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+                                tags=self.CHECK_TAGS, at_least=1)
 
         self.coverage_report()
 
@@ -514,7 +494,7 @@ class SNMPTestCase(AgentCheckTest):
 
 
         self.run_check_n(config, repeat=3, sleep=2)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for symbol in self.TABULAR_OBJECTS[0]['symbols']:
@@ -528,7 +508,7 @@ class SNMPTestCase(AgentCheckTest):
 
         # Test service check
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+                                tags=self.CHECK_TAGS, at_least=1)
 
         self.coverage_report()
 
@@ -542,12 +522,12 @@ class SNMPTestCase(AgentCheckTest):
         }
         self.run_check(config)
 
-        self.warnings = self.wait_for_async('get_warnings', 'warnings', 1)
+        self.warnings = self.wait_for_async('get_warnings', 'warnings', 1, RESULTS_TIMEOUT)
         self.assertWarning("Fail to collect some metrics: No symbol IF-MIB::noIdeaWhatIAmDoingHere",
                            count=1, exact_match=False)
 
         # # Test service check
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.CRITICAL,
                                 tags=self.CHECK_TAGS, count=1)
         self.coverage_report()
@@ -560,7 +540,7 @@ class SNMPTestCase(AgentCheckTest):
             'instances': [self.generate_instance_config(self.FORCED_METRICS)]
         }
         self.run_check_twice(config)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         for metric in self.FORCED_METRICS:
             metric_name = "snmp." + (metric.get('name') or metric.get('symbol'))
@@ -574,7 +554,7 @@ class SNMPTestCase(AgentCheckTest):
 
         # # Test service check
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+                                tags=self.CHECK_TAGS, at_least=1)
         self.coverage_report()
 
     def test_invalid_forcedtype_metric(self):
@@ -588,11 +568,11 @@ class SNMPTestCase(AgentCheckTest):
 
         self.run_check(config)
 
-        self.warnings = self.wait_for_async('get_warnings', 'warnings', 1)
+        self.warnings = self.wait_for_async('get_warnings', 'warnings', 1, RESULTS_TIMEOUT)
         self.assertWarning("Invalid forced-type specified:", count=1, exact_match=False)
 
         # # Test service check
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.CRITICAL,
                                 tags=self.CHECK_TAGS, count=1)
         self.coverage_report()
@@ -605,7 +585,7 @@ class SNMPTestCase(AgentCheckTest):
             'instances': [self.generate_instance_config(self.SCALAR_OBJECTS_WITH_TAGS)]
         }
         self.run_check_n(config, repeat=3)
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
 
         # Test metrics
         for metric in self.SCALAR_OBJECTS_WITH_TAGS:
@@ -613,8 +593,7 @@ class SNMPTestCase(AgentCheckTest):
             tags = self.CHECK_TAGS + metric.get('metric_tags')
             self.assertMetric(metric_name, tags=tags, count=1)
         # Test service check
-        self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK,
-                                tags=self.CHECK_TAGS, count=1)
+        self.assertServiceCheck("snmp.can_check", status=AgentCheck.OK, tags=self.CHECK_TAGS, at_least=1)
 
         self.coverage_report()
 
@@ -631,12 +610,12 @@ class SNMPTestCase(AgentCheckTest):
             'instances': [instance]
         }
         self.run_check(config)
-        self.warnings = self.wait_for_async('get_warnings', 'warnings', 1)
+        self.warnings = self.wait_for_async('get_warnings', 'warnings', 1, RESULTS_TIMEOUT)
 
         self.assertWarning("No SNMP response received before timeout for instance localhost", count=1)
 
         # Test service check
-        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1)
+        self.service_checks = self.wait_for_async('get_service_checks', 'service_checks', 1, RESULTS_TIMEOUT)
         self.assertServiceCheck("snmp.can_check", status=AgentCheck.CRITICAL,
                                 tags=self.CHECK_TAGS, count=1)