+++ docker_daemon/CHANGELOG.md
@@ -1,10 +1,25 @@
 # CHANGELOG - docker_daemon
 
-1.6.0 / Unreleased
+1.7.0 / Unreleased
 ==================
 ### Changes
-* [BUGFIX] Fix lost kubernetes tags in 1.5.0. See [#817][]
+
+* [FEATURE] Honor global collect_labels_as_tags if integration's collect_labels_as_tags is empty. See [#881][]
+* [IMPROVEMENT] Improve logging when cgroup metrics can't be retrieved. See [#914][]
+
+1.6.0 / 2017-11-21
+==================
+### Changes
+
 * [IMPROVEMENT] Add custom tags to all service checks. See [#782][]
+* [IMPROVEMENT] Add docker memory soft limit metric. See [#760][]
+* [IMPROVEMENT] Add docker.containers.running.total & docker.containers.stopped.total metrics. See [#859][]
+
+1.5.1 / 2017-11-08
+==================
+### Changes
+
+* [BUGFIX] Fix lost kubernetes tags in 1.5.0. See [#817][]
 
 1.5.0 / 2017-10-10
 ==================
@@ -107,6 +122,9 @@
 [#701]: https://github.com/DataDog/integrations-core/issues/701
 [#722]: https://github.com/DataDog/integrations-core/issues/722
 [#744]: https://github.com/DataDog/integrations-core/issues/744
+[#760]: https://github.com/DataDog/integrations-core/issues/760
 [#770]: https://github.com/DataDog/integrations-core/issues/770
 [#782]: https://github.com/DataDog/integrations-core/issues/782
+[#817]: https://github.com/DataDog/integrations-core/issues/817
+[#859]: https://github.com/DataDog/integrations-core/issues/859
 [@sophaskins]: https://github.com/sophaskins